+++ src/app/done/done.component.ts
@@ -1,4 +1,6 @@
 import { Component, OnInit } from '@angular/core';
+import { FormGroup, FormControl, Validators } from '@angular/forms';
+import { NgxSmartModalService } from 'ngx-smart-modal';
 import { Card } from '../model/card';
 import { DoneService } from './done.service';
 
@@ -10,8 +12,15 @@ import { DoneService } from './done.service';
 export class DoneComponent implements OnInit {
 
   dones: Card[];
+  editForm = new FormGroup({
+    id: new FormControl(),
+    title: new FormControl('', Validators.required),
+    description: new FormControl('', Validators.required),
+    status: new FormControl('TODO')
+  });
 
-  constructor(private doneService: DoneService) {
+  constructor(private doneService: DoneService, 
+        private modalService: NgxSmartModalService) {
   }
 
   ngOnInit(): void {
@@ -25,9 +34,35 @@ export class DoneComponent implements OnInit {
   }
 
   removeCard(card: Card){
-   this.doneService.deleteDone(card).subscribe( done => {
-     this.getCardsDones();
-   });
+    if(confirm("Deseja relmente deletar esse card?")){
+      this.doneService.deleteDone(card).subscribe( done => {
+        this.getCardsDones();
+      });
+    }
+  }
+
+  edit(card: Card) {
+    this.editForm.patchValue({
+      id: card.id,
+      title: card.title,
+      description: card.description,
+      status: card.status
+    });
+    this.modalService.getModal("cardEditModalDone").open();
+  }
+
+  onSubmit() {
+    console.log(this.editForm.value);
+    this.doneService.updateCard(this.editForm.value).subscribe(card => {
+      this.editForm.patchValue({
+        id: card.id,
+        title: card.title,
+        description: card.description,
+        status: card.status
+      });
+    });
+    this.modalService.getModal("cardEditModalDone").close();
+    location.reload();
   }
 
 }
