+++ src/app/todo/todo.component.ts
@@ -1,4 +1,6 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, EventEmitter, OnInit, Output } from '@angular/core';
+import { FormGroup, FormControl, Validators } from '@angular/forms';
+import { NgxSmartModalService } from 'ngx-smart-modal';
 import { Card } from '../model/card';
 import { TodoService } from './todo.service';
 
@@ -9,9 +11,19 @@ import { TodoService } from './todo.service';
 })
 export class TodoComponent implements OnInit {
 
-  todos: Card[];
+  editForm = new FormGroup({
+    id: new FormControl(),
+    title: new FormControl('', Validators.required),
+    description: new FormControl('', Validators.required),
+    status: new FormControl('TODO')
+  });
 
-  constructor(private todoService: TodoService) { }
+  todos: Card[];
+  @Output() currentTodo = new EventEmitter();
+  @Output() isEdit = new EventEmitter();
+
+  constructor(private todoService: TodoService,
+          private modalService: NgxSmartModalService) { }
 
   ngOnInit(): void {
     this.getTodos();
@@ -24,9 +36,34 @@ export class TodoComponent implements OnInit {
   }
 
   removeTodo(card: Card) {
-    this.todoService.deleteTodo(card).subscribe( todo => {
-      this.getTodos();
-    })
+    if(confirm("Deseja relmente deletar esse card?")){
+      this.todoService.deleteTodo(card).subscribe( todo => {
+        this.getTodos();
+      })
+    }
   }
 
+  edit(card: Card) {
+    this.editForm.patchValue({
+      id: card.id,
+      title: card.title,
+      description: card.description,
+      status: card.status
+    });
+    this.modalService.getModal("cardEditModal").open();
+  }
+
+  onSubmit() {
+    console.log(this.editForm.value);
+    this.todoService.updateCard(this.editForm.value).subscribe(card => {
+      this.editForm.patchValue({
+        id: card.id,
+        title: card.title,
+        description: card.description,
+        status: card.status
+      });
+    });
+    this.modalService.getModal("cardEditModal").close();
+    location.reload();
+  }
 }
